Fix img width and affix. Closes #94
This commit fixes the affix issue with the image. The main problem is the width
and height of the image, and obviously the browser window of the user.

In order to keep it save, I've used a small width and height, keeping the
aspect ration, as the volunteer can zoom in/out.